LPS-195103 Move SqlUpdateImpl and MappingSqlQueryImpl to kernel prepping next
