+++ portal-kernel/src/com/liferay/portal/kernel/dao/jdbc/MappingSqlQueryImpl.java
@@ -3,15 +3,18 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.dao.jdbc.spring;
+package com.liferay.portal.kernel.dao.jdbc;
 
 import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.kernel.dao.jdbc.ParamSetter;
-import com.liferay.portal.kernel.dao.jdbc.SqlUpdate;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 import javax.sql.DataSource;
 
@@ -19,18 +22,20 @@ import javax.sql.DataSource;
  * @author Brian Wing Shun Chan
  * @author Shuyang Zhou
  */
-public class SqlUpdateImpl implements SqlUpdate {
+public class MappingSqlQueryImpl<T> implements MappingSqlQuery<T> {
 
-	public SqlUpdateImpl(
-		DataSource dataSource, String sql, ParamSetter... paramSetters) {
+	public MappingSqlQueryImpl(
+		DataSource dataSource, String sql, RowMapper<T> rowMapper,
+		ParamSetter... paramSetters) {
 
 		_dataSource = dataSource;
 		_sql = sql;
+		_rowMapper = rowMapper;
 		_paramSetters = paramSetters;
 	}
 
 	@Override
-	public int update(Object... params) throws SQLException {
+	public List<T> execute(Object... params) throws SQLException {
 		if (_paramSetters.length != params.length) {
 			throw new IllegalArgumentException(
 				StringBundler.concat(
@@ -48,12 +53,29 @@ public class SqlUpdateImpl implements SqlUpdate {
 				paramSetter.set(preparedStatement, i + 1, params[i]);
 			}
 
-			return preparedStatement.executeUpdate();
+			List<T> results = null;
+
+			try (ResultSet resultSet = preparedStatement.executeQuery()) {
+				while (resultSet.next()) {
+					if (results == null) {
+						results = new ArrayList<>();
+					}
+
+					results.add(_rowMapper.mapRow(resultSet));
+				}
+			}
+
+			if (results == null) {
+				results = Collections.emptyList();
+			}
+
+			return results;
 		}
 	}
 
 	private final DataSource _dataSource;
 	private final ParamSetter[] _paramSetters;
+	private final RowMapper<T> _rowMapper;
 	private final String _sql;
 
 }
